backend/OCRService/requirements.txt
fastapi
uvicorn
pydantic
psycopg2
sqlalchemy
pillow
pytesseract
python-dotenv
